requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
python-dateutil==2.8.2
Werkzeug==3.0.1
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
